[WebAssembly(WIP)] Add location relative relocation for table index
This is a variant of aa0c571a5fa9130ccd932f9b6970440a1bae177f for table
index. This change is required for representing relative function
pointer in Swift runtime metadata.

But Wasm is a harvard architecture, and its data pointer and function
pointer have different address spaces. So relative function pointer is
not semantically relative, but just a subtraction relocation.